backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Talib1073", email = "talib.hussain1934@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "cohere>=5.20.1",
    "fastapi>=0.115.0",
    "lxml>=6.0.2",
    "openai-agents>=0.6.7",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.2",
    "requests>=2.32.5",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "uvicorn>=0.32.0",
]

[build-system]
requires = ["uv_build>=0.9.25,<0.10.0"]
build-backend = "uv_build"
